added renamed Buffer to DataBuffer made the switched to Byte/byte where I could for consistancy to spec

+++ src/tbas/DataTape.java
@@ -3,24 +3,24 @@ package tbas;
 import java.util.ArrayList;
 
 public class DataTape {
-	private ArrayList<Integer> dataTape;
+	private ArrayList<Byte> dataTape;
 	private int dataPointer;
 
 	public DataTape() {
-		dataTape = new ArrayList<Integer>();
-		dataTape.add(0);
+		dataTape = new ArrayList<Byte>();
+		dataTape.add((byte) 0);
 		dataPointer = 0;
 	}
 
 	public void incrementCellData() {
 		if (dataTape.get(dataPointer) < 255) {
-			dataTape.set(dataPointer, dataTape.get(dataPointer) + 1);
+			dataTape.set(dataPointer,(byte) (dataTape.get(dataPointer) + 1));
 		}
 	}
 
 	public void decrementCellData() {
 		if (dataTape.get(dataPointer) > 0) {
-			dataTape.set(dataPointer, dataTape.get(dataPointer) - 1);
+			dataTape.set(dataPointer, (byte) (dataTape.get(dataPointer) - 1));
 		}
 	}
 
@@ -31,7 +31,7 @@ public class DataTape {
 	public void advanceTape() {
 		dataPointer += 1;
 		if (dataTape.size() - 1 < dataPointer) {
-			dataTape.add(0);
+			dataTape.add((byte)0);
 		}
 	}
 
@@ -43,15 +43,15 @@ public class DataTape {
 		dataPointer += offset;
 	}
 
-	public Integer readTape() {
-		return dataTape.get(dataPointer); 
+	public Byte readTape() {
+		return dataTape.get(dataPointer);
 	}
 
-	public ArrayList<Integer> getTapeData() {
+	public ArrayList<Byte> getTapeData() {
 		return dataTape;
 	}
 
-	public void setTapeData(ArrayList<Integer> tapeData) {
+	public void setTapeData(ArrayList<Byte> tapeData) {
 		this.dataTape = tapeData;
 	}
 
@@ -63,8 +63,8 @@ public class DataTape {
 		this.dataPointer = dataPointer;
 	}
 
-	public void write(int input) {
-		if(input < 255){
+	public void write(Byte input) {
+		if (input < 255) {
 			dataTape.set(dataPointer, input);
 		}
 	}
